refactor: Same models for both MeteringPoint clients
