+++ domains/certificates/Query.API/API.UnitTests/Query.API/ApiModels/Responses/MeteringPointTests.cs
@@ -1,5 +1,5 @@
-using API.ContractService.Clients;
 using Xunit;
+using API.ContractService.Models;
 
 namespace API.UnitTests.Query.API.ApiModels.Responses;
