+++ domains/certificates/Query.API/API/ContractService/Clients/Internal/AdminPortalMeteringPointsClient.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Net.Http.Json;
@@ -6,6 +5,7 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Threading;
 using System.Threading.Tasks;
+using API.ContractService.Models;
 using Microsoft.AspNetCore.Http;
 
 namespace API.ContractService.Clients.Internal;
@@ -43,41 +43,3 @@ public class AdminPortalMeteringPointsClient(HttpClient httpClient, IHttpContext
             AuthenticationHeaderValue.Parse(httpContextAccessor.HttpContext!.Request.Headers.Authorization!);
     }
 }
-
-public record MeteringPointsResponse(List<MeteringPoint> Result);
-
-public record MeteringPoint(
-    string Gsrn,
-    string GridArea,
-    MeterType MeteringPointType,
-    SubMeterType SubMeterType,
-    Address Address,
-    Technology Technology,
-    string ConsumerCvr,
-    bool CanBeUsedForIssuingCertificates,
-    string Capacity,
-    string BiddingZone);
-
-public enum MeterType
-{
-    Consumption,
-    Production,
-    Child
-}
-
-public enum SubMeterType
-{
-    Physical,
-    Virtual,
-    Calculated
-}
-
-public record Technology(string AibFuelCode, string AibTechCode);
-
-public record Address(
-    string Address1,
-    string? Address2,
-    string? Locality,
-    string City,
-    string PostalCode,
-    string Country);
