+++ domains/certificates/Query.API/API/ContractService/Clients/Internal/IAdminPortalMeteringPointsClient.cs
@@ -1,5 +1,6 @@
 using System.Threading;
 using System.Threading.Tasks;
+using API.ContractService.Models;
 
 namespace API.ContractService.Clients.Internal;
