+++ domains/certificates/Query.API/API/ContractService/Clients/MeteringPointsClient.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Net.Http.Json;
@@ -7,6 +6,7 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Threading;
 using System.Threading.Tasks;
+using API.ContractService.Models;
 using EnergyOrigin.Setup;
 using EnergyOrigin.TokenValidation.b2c;
 using Microsoft.AspNetCore.Http;
@@ -75,41 +75,3 @@ public class MeteringPointsClient : IMeteringPointsClient
         }
     }
 }
-
-public record MeteringPointsResponse(List<MeteringPoint> Result);
-
-public record MeteringPoint(
-    string Gsrn,
-    string GridArea,
-    MeterType MeteringPointType,
-    SubMeterType SubMeterType,
-    Address Address,
-    Technology Technology,
-    string ConsumerCvr,
-    bool CanBeUsedForIssuingCertificates,
-    string Capacity,
-    string BiddingZone);
-
-public enum MeterType
-{
-    Consumption,
-    Production,
-    Child
-}
-
-public enum SubMeterType
-{
-    Physical,
-    Virtual,
-    Calculated
-}
-
-public record Technology(string AibFuelCode, string AibTechCode);
-
-public record Address(
-    string Address1,
-    string? Address2,
-    string? Locality,
-    string City,
-    string PostalCode,
-    string Country);
