+++ domains/certificates/Query.API/API/ContractService/ContractServiceImpl.cs
@@ -151,17 +151,17 @@ internal class ContractServiceImpl : IContractService
         return contracts.ToList();
     }
 
-    private static MeteringPointType Map(MeterType type)
+    private static MeteringPointType Map(Models.MeterType type)
     {
-        if (type == MeterType.Production) return MeteringPointType.Production;
-        if (type == MeterType.Consumption) return MeteringPointType.Consumption;
+        if (type == Models.MeterType.Production) return MeteringPointType.Production;
+        if (type == Models.MeterType.Consumption) return MeteringPointType.Consumption;
 
         throw new ArgumentException($"Unsupported MeterType {type}");
     }
 
-    private static Technology? Map(MeterType meterType, Clients.Technology technology)
+    private static Technology? Map(Models.MeterType meterType, Models.Technology technology)
     {
-        if (meterType == MeterType.Production)
+        if (meterType == Models.MeterType.Production)
         {
             return new Technology(technology.AibFuelCode, technology.AibTechCode);
         }
